--- a/src/main/java/org/ilastik/ilastik4ij/IlastikExport.java
+++ b/src/main/java/org/ilastik/ilastik4ij/IlastikExport.java
@@ -6,6 +6,8 @@
 package org.ilastik.ilastik4ij;
 
 import java.io.File;
+
+import ij.IJ;
 import net.imagej.DatasetService;
 import org.scijava.app.StatusService;
 import org.scijava.command.Command;
@@ -18,47 +20,47 @@ import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetWriter;
 import org.scijava.ItemVisibility;
 
 /**
- *
  * @author chaubold
  */
 @Plugin(type = Command.class, headless = false, menuPath = "Plugins>ilastik>Export HDF5")
 public class IlastikExport implements Command {
-
-    // needed services:
     @Parameter
-    LogService log;
-
+    private LogService log;
     @Parameter
-    StatusService statusService;
-
+    private StatusService statusService;
     @Parameter
-    DatasetService datasetService;
-    
+    private DatasetService datasetService;
+
     @Parameter
     OptionsService optionsService;
 
     // plugin parameters
     @Parameter(visibility = ItemVisibility.MESSAGE)
     private String message = "Be careful to properly configure axes.\nScrollabale axes"
-					+ " time frames, channels and z-slices can be easily mistaken."
-                    + "\nFor changing their order go to"
-					+ " Image -> Properties";
-    
-    @Parameter(label = "HDF5 file where to export to", style="save")
+            + " time frames, channels and z-slices can be easily mistaken."
+            + "\nFor changing their order go to"
+            + " Image -> Properties";
+
+    @Parameter(label = "HDF5 file where to export to", style = "save")
     private File hdf5FileName;
-    
+
     @Parameter(label = "Image to save")
     private Dataset input;
 
-    @Parameter(label = "Compression level (0-9)", style="spinner", min = "0", max="9",
-               description = "The best setting depends on the kind of data you are saving."
-                             + " Segmentations can be compressed well (-> select 9), "
-                             + "but raw data is best saved without compression (->0) for faster access.")
+    @Parameter(label = "Compression level (0-9)", style = "spinner", min = "0", max = "9",
+            description = "The best setting depends on the kind of data you are saving."
+                    + " Segmentations can be compressed well (-> select 9), "
+                    + "but raw data is best saved without compression (->0) for faster access.")
     private int compressionLevel = 0;
-    
+
     @Override
     public void run() {
-        new Hdf5DataSetWriter(input.getImgPlus(), hdf5FileName.getAbsolutePath(), "data", compressionLevel, log, statusService).write();
+        String filename = hdf5FileName.getAbsolutePath();
+        if (!filename.endsWith(".h5")) {
+            IJ.error("Error: HDF5 export file must have '.h5' suffix");
+        } else {
+            new Hdf5DataSetWriter(input.getImgPlus(), filename, "data", compressionLevel, log, statusService).write();
+        }
     }
-    
+
 }