--- a/src/main/java/org/ilastik/ilastik4ij/IlastikImport.java
+++ b/src/main/java/org/ilastik/ilastik4ij/IlastikImport.java
@@ -10,22 +10,6 @@ import ch.systemsx.cisd.hdf5.HDF5Factory;
 import ch.systemsx.cisd.hdf5.HDF5LinkInformation;
 import ch.systemsx.cisd.hdf5.IHDF5Reader;
 import ij.IJ;
-import java.awt.BorderLayout;
-import java.awt.GridBagConstraints;
-import java.awt.GridBagLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
-import javax.swing.JButton;
-import javax.swing.JComboBox;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import ncsa.hdf.hdf5lib.exceptions.HDF5Exception;
 import net.imagej.DatasetService;
 import net.imagej.ImgPlus;
 import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetReader;
@@ -38,12 +22,31 @@ import org.scijava.options.OptionsService;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.io.File;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
+
 /**
- *
  * @author chaubold
  */
 @Plugin(type = Command.class, headless = false, menuPath = "Plugins>ilastik>Import HDF5")
 public class IlastikImport implements Command, ActionListener {
+    private static final String SELECT_DATASET = "selectDataset";
+    private static final String CANCEL_DATASET_SELECTION = "cancelDatasetSelection";
+    private static final String LOAD_RAW = "loadRaw";
+    private static final String LOAD_LUT = "loadLUT";
+    private static final String CANCEL_AXES_ORDER_CONFIGURATION = "cancelAxesOrderConfiguration";
 
     // needed services:
     @Parameter
@@ -51,14 +54,14 @@ public class IlastikImport implements Command, ActionListener {
 
     @Parameter
     DatasetService datasetService;
-    
+
     @Parameter
     OptionsService optionsService;
 
     // plugin parameters
     @Parameter(label = "HDF5 file exported from ilastik")
     private File hdf5FileName;
-    
+
     @Parameter(type = ItemIO.OUTPUT)
     private ImgPlus output;
 
@@ -73,7 +76,7 @@ public class IlastikImport implements Command, ActionListener {
     private JFrame frameSelectDataset;
     private String dimensionOrder;
     private String datasetPath;
-    
+
     private final Lock lock = new ReentrantLock();
     private final Condition finishedCondition = lock.newCondition();
     private boolean isFinished = false;
@@ -94,25 +97,23 @@ public class IlastikImport implements Command, ActionListener {
                 this.isList = true;
             }
 
-        } catch (HDF5Exception err) {
-            IJ.error("Error while opening '" + fullFileName + err);
         } catch (Exception err) {
             IJ.error("Error while opening '" + fullFileName + err);
         } catch (OutOfMemoryError o) {
             IJ.outOfMemory("Load HDF5");
         }
-        
+
         // wait for isFinished to become true
         lock.lock();
         try {
-            while(!isFinished)
+            while (!isFinished)
                 finishedCondition.await();
         } catch (InterruptedException ex) {
             log.warn("Execution of HDF5 loading got interrupted");
         } finally {
             lock.unlock();
         }
-        
+
         log.info("Done loading HDF5 file!");
     }
 
@@ -143,47 +144,33 @@ public class IlastikImport implements Command, ActionListener {
 
     private void showAxesorderInputDialog() {
         String boxInfo;
-        int rank = 0;
         String[] dimExamples = new String[20];
 
         frameSelectAxisOrdering = new JFrame();
 
-        JLabel datasetLabel = new JLabel(); 
+        JLabel datasetLabel = new JLabel();
         JLabel taskLabel = new JLabel("Please enter the meaning of those axes:");
         JButton k2 = new JButton("Cancel");
-        k2.setActionCommand("cancelAxesOrderConfiguration");
+        k2.setActionCommand(CANCEL_AXES_ORDER_CONFIGURATION);
         k2.addActionListener(this);
 
         if (this.isList) {
             boxInfo = (String) dataSetBox.getSelectedItem();
             String[] parts = boxInfo.split(":");
             datasetPath = parts[1].replaceAll("\\s+", "");
-//			log.info(boxInfo);
         } else {
             datasetPath = datasetList.get(0);
-//			log.info(path);
         }
         HDF5DataSetInformation dsInfo = reader.object().getDataSetInformation(datasetPath);
-        rank = dsInfo.getRank();
-        
-        String datasetDescription = "Found dataset with dimensions: (";
-        long[] dims = dsInfo.getDimensions();
-        boolean isFirst = true;
-        for(long dim : dims)
-        {
-            if(isFirst)
-            {
-                datasetDescription = datasetDescription + String.valueOf(dim);
-                isFirst = false;
-            }
-            else
-                datasetDescription = datasetDescription + ", " + String.valueOf(dim);
-        }
-        
-        datasetDescription = datasetDescription + ")";
+
+        String shape = Arrays.stream(dsInfo.getDimensions())
+                .mapToObj(String::valueOf)
+                .collect(Collectors.joining(", "));
+
+        String datasetDescription = String.format("Found dataset with dimensions: (%s)", shape);
         datasetLabel.setText(datasetDescription);
 
-        switch (rank) {
+        switch (dsInfo.getRank()) {
             case 5:
                 dimExamples[0] = "tzyxc";
                 dimExamples[1] = "txyzc";
@@ -207,30 +194,30 @@ public class IlastikImport implements Command, ActionListener {
         this.dimBox = new JComboBox(dimExamples);
         dimBox.setEditable(true);
         dimBox.addActionListener(this);
-        
+
         JButton l1 = new JButton("Load Raw");
-        l1.setActionCommand("Load Raw");
+        l1.setActionCommand(LOAD_RAW);
         l1.addActionListener(this);
         JButton l2 = new JButton("Load and apply LUT");
-        l2.setActionCommand("Load LUT");
+        l2.setActionCommand(LOAD_LUT);
         l2.addActionListener(this);
 
         // layout frame:
         frameSelectAxisOrdering.getContentPane().setLayout(new GridBagLayout());
         GridBagConstraints c = new GridBagConstraints();
-        
+
         c.fill = GridBagConstraints.HORIZONTAL;
         c.gridx = 0;
         c.gridy = 0;
         c.gridwidth = 3;
         frameSelectAxisOrdering.getContentPane().add(datasetLabel, c);
-        
+
         c.fill = GridBagConstraints.HORIZONTAL;
         c.gridx = 0;
         c.gridy = 1;
         c.gridwidth = 3;
         frameSelectAxisOrdering.getContentPane().add(taskLabel, c);
-        
+
         c.fill = GridBagConstraints.HORIZONTAL;
         c.gridx = 0;
         c.gridy = 2;
@@ -242,7 +229,7 @@ public class IlastikImport implements Command, ActionListener {
         c.gridy = 3;
         c.gridwidth = 1;
         frameSelectAxisOrdering.getContentPane().add(l1, c);
-        
+
         c.fill = GridBagConstraints.HORIZONTAL;
         c.gridx = 1;
         c.gridy = 3;
@@ -254,7 +241,7 @@ public class IlastikImport implements Command, ActionListener {
         c.gridy = 3;
         c.gridwidth = 1;
         frameSelectAxisOrdering.getContentPane().add(k2, c);
-        
+
         frameSelectAxisOrdering.setResizable(false);
         frameSelectAxisOrdering.setLocationRelativeTo(null);
         frameSelectAxisOrdering.pack();
@@ -264,10 +251,10 @@ public class IlastikImport implements Command, ActionListener {
     private void showDatasetSelectionDialog(IHDF5Reader reader, List<String> datasetList) {
         frameSelectDataset = new JFrame();
         JButton b1 = new JButton("Select");
-        b1.setActionCommand("selectDataset");
+        b1.setActionCommand(SELECT_DATASET);
         b1.addActionListener(this);
         JButton b2 = new JButton("Cancel");
-        b2.setActionCommand("cancelDatasetSelection");
+        b2.setActionCommand(CANCEL_DATASET_SELECTION);
         b2.addActionListener(this);
 
         String[] dataSets = new String[datasetList.size()];
@@ -300,38 +287,52 @@ public class IlastikImport implements Command, ActionListener {
 
     }
 
-    private void signalFinished(){
+    private void signalFinished() {
         lock.lock();
-        try{
+        try {
             isFinished = true;
             finishedCondition.signal();
-        } finally
-        {
+        } finally {
             lock.unlock();
         }
     }
-    
+
     @Override
     public void actionPerformed(ActionEvent event) {
-        if (event.getActionCommand().equals("selectDataset")) {
-            showAxesorderInputDialog();
-        } else if (event.getActionCommand().equals("cancelDatasetSelection")) {
-            frameSelectDataset.dispose();
-            signalFinished();
-        } else if (event.getActionCommand().equals("Load Raw")) {
-            dimensionOrder = (String) dimBox.getSelectedItem();
-            frameSelectAxisOrdering.dispose();
-            output = new Hdf5DataSetReader(fullFileName, datasetPath, dimensionOrder, log, datasetService).read();
-            signalFinished();
-        } else if (event.getActionCommand().equals("Load LUT")) {
-            dimensionOrder = (String) dimBox.getSelectedItem();
-            frameSelectAxisOrdering.dispose();
-            output = new Hdf5DataSetReader(fullFileName, datasetPath, dimensionOrder, log, datasetService).read();
-            IJ.run("3-3-2 RGB"); // Applies the lookup table
-            signalFinished();
-        } else if (event.getActionCommand().equals("cancelAxesOrderConfiguration")) {
-            frameSelectAxisOrdering.dispose();
-            signalFinished();
+        String actionCommand = event.getActionCommand();
+        switch (actionCommand) {
+            case SELECT_DATASET:
+                frameSelectDataset.dispose();
+                showAxesorderInputDialog();
+                break;
+            case CANCEL_DATASET_SELECTION:
+                frameSelectDataset.dispose();
+                signalFinished();
+                break;
+            case LOAD_RAW:
+                dimensionOrder = (String) dimBox.getSelectedItem();
+                frameSelectAxisOrdering.dispose();
+                Instant start = Instant.now();
+
+                output = new Hdf5DataSetReader(fullFileName, datasetPath, dimensionOrder, log, datasetService).read();
+
+                Instant finish = Instant.now();
+                long timeElapsed = Duration.between(start, finish).toMillis();
+                log.info("Loading HDF5 dataset took: " + timeElapsed);
+
+                signalFinished();
+                break;
+            case LOAD_LUT:
+                dimensionOrder = (String) dimBox.getSelectedItem();
+                frameSelectAxisOrdering.dispose();
+                output = new Hdf5DataSetReader(fullFileName, datasetPath, dimensionOrder, log, datasetService).read();
+                IJ.run("3-3-2 RGB"); // Applies the lookup table
+                signalFinished();
+                break;
+            case CANCEL_AXES_ORDER_CONFIGURATION:
+                frameSelectAxisOrdering.dispose();
+                signalFinished();
+                break;
         }
     }