… pom.xml
… src/main/java/org/ilastik/ilastik4ij/IlastikObjectClassificationPrediction.java
… src/main/java/org/ilastik/ilastik4ij/IlastikPixelClassificationPrediction.java
… src/main/java/org/ilastik/ilastik4ij/IlastikTrackingPrediction.java
… src/main/java/org/ilastik/ilastik4ij/hdf5/Hdf5DataSetConfig.java
… src/main/java/org/ilastik/ilastik4ij/hdf5/Hdf5DataSetReader.java
… src/main/java/org/ilastik/ilastik4ij/hdf5/Hdf5DataSetWriter.java
… src/main/java/org/ilastik/ilastik4ij/util/ComboBoxDimensions.java
… src/main/java/org/ilastik/ilastik4ij/util/DatasetInfo.java
… src/main/java/org/ilastik/ilastik4ij/util/Hdf5Utils.java
… src/main/java/org/ilastik/ilastik4ij/util/IlastikBoxModel.java
… src/test/java/org/ilastik/ilastik4ij/Hdf5DataSetReaderTest.java
… src/test/resources/2d_cells_apoptotic_1channel.h5